LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/adoption.window-expected.txt


PASS Adoption with global registry
PASS Adoption with explicit global registry
PASS Adoption with scoped registry
PASS Adoption with global registry into a scoped registry
PASS Adoption with explicit global registry into a scoped registry
PASS Adoption with scoped registry into a scoped registry
FAIL Adoption including shadow root with global registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry
FAIL Adoption including shadow root with global registry into a scoped registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry into a scoped registry assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry into a scoped registry
PASS Adoption with global registry (null registry target)
PASS Adoption with explicit global registry (null registry target)
PASS Adoption with scoped registry (null registry target)
PASS Adoption with global registry into a scoped registry (null registry target)
PASS Adoption with explicit global registry into a scoped registry (null registry target)
PASS Adoption with scoped registry into a scoped registry (null registry target)
FAIL Adoption including shadow root with global registry (null registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry (null registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry (null registry target)
FAIL Adoption including shadow root with global registry into a scoped registry (null registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry into a scoped registry (null registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry into a scoped registry (null registry target)
FAIL Adoption with global registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption with explicit global registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption with scoped registry (scoped registry target)
FAIL Adoption with global registry into a scoped registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption with explicit global registry into a scoped registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption with scoped registry into a scoped registry (scoped registry target)
FAIL Adoption including shadow root with global registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry (scoped registry target)
FAIL Adoption including shadow root with global registry into a scoped registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL Adoption including shadow root with explicit global registry into a scoped registry (scoped registry target) assert_equals: expected null but got object "[object CustomElementRegistry]"
PASS Adoption including shadow root with scoped registry into a scoped registry (scoped registry target)
